======================================================================================= This tutorial involved understanding key concepts in OpenGL, including VAOs (Vertex Attribute Objects), VBOs (Vertex Buffer Objects), and EBOs (Element Buffer Objects). The first end result was drawing a simple filled triangle on the screen, which involved creating a VAO and VBO. This involved understanding how binding works, how loading memory into VRAM is done, and understanding how OpenGL functions essentially like a giant "State Machine" (in which a VAO is almost like a 'state capture').
The final goal was to draw a rectangle to the screen, and this would naively be done by drawing two separate triangles––however, this incurs repetition in the `vertices[]` array. To avoid this, we make use of an Element Buffer Object (which is created and loaded very similarly to a VBO), and then specify the *order* in which we would like to draw the vertices. This means we only need to store **unique** vertices, not repeated ones.

Finally, this tutorial introduced the concept of shaders, and in particular that OpenGL requires that we *at least* define our own **vertex** and **fragment** shaders. These were both done using string literals inside of `main.cpp`, and the vertex shader simply forwarded vertex position info, while the fragment shader simply colored every fragment a predefined shade of orange.

In the end, I learned a lot about the purpose and basic functionality of VAOs, VBOs, and EBOs. I also learned how to write the simplest versions of vertex and fragment shaders, as well as create a program, attach the shaders to them, and then actually *use* that program when rendering.